fix(gameclient): widen the listbox scroll offset past a signed short
ListboxData::displayPos and displayHeight were Short while the pixel
totals they are assigned from and compared against, ListEntryRow::
listHeight and ListboxData::totalHeight, are Int. Past roughly 2185
rows the assignment in GadgetListBoxSetTopVisibleEntry and the arrow,
slider and wheel paths wrapped negative, so the view jumped and the
"displayPos > 0" guard then refused to scroll back up. Reachable from
the replay list, the map list and a long chat buffer.

Ported by hand from a known upstream fix.

Witness: listbox_scroll_offset_survives_a_list_taller_than_a_signed_short
in test_gameengine. Non-vacuous: with displayPos back to Short, 4 of its
5 checks fail.

+++ GeneralsMD/Code/GameEngine/Include/GameClient/Gadget.h
@@ -377,11 +377,15 @@ typedef struct _ListboxData
 	Int					selectPos;				// Position of current selected entry (for SINGLE select)
 	Int					*selections;			// Pointer to array of selections (for MULTI select)
 
-	Short				displayHeight;		// Height in pixels of listbox display region
+	Int					displayHeight;		// Height in pixels of listbox display region
 																// this is computed based on the existance
 																// of a title or not.
 	UnsignedInt doubleClickTime;	//
-	Short				displayPos;				// Position of current display entry in pixels
+	// both of these are pixel offsets into totalHeight, which is an Int, and displayPos is assigned
+	// straight out of ListEntryRow::listHeight.  A list whose entries add up to more than 32767
+	// pixels - a few thousand chat lines or replays - used to wrap these to a negative number and
+	// the scrollbar jumped to the wrong end of the list.
+	Int					displayPos;				// Position of current display entry in pixels
 
 } ListboxData;
